+++ cmd/root.go
@@ -328,10 +328,10 @@ func AddFlags(flagSet *pflag.FlagSet, cfg *config.Config) (configPath *string) {
 	flagSet.DurationVar(&cfg.POET.SubmitChallengeTimeout, "submit-challenge-timeout",
 		cfg.POET.SubmitChallengeTimeout,
 		"time within grace period, when post challenge must be submitted to poet server")
-	flagSet.DurationVar(&cfg.POET.GetProofTimeout, "get-proof-timeout",
-		cfg.POET.GetProofTimeout, "time period when ready poet proof must be fetched")
-	flagSet.DurationVar(&cfg.POET.DefaultRequestTimeout, "poet-default-request-timeout",
-		cfg.POET.DefaultRequestTimeout, "default timeout for poet requests")
+	flagSet.DurationVar(&cfg.POET.FetchProofTimeout, "get-proof-timeout",
+		cfg.POET.FetchProofTimeout, "time period when ready poet proof must be fetched")
+	flagSet.DurationVar(&cfg.POET.RequestTimeout, "poet-default-request-timeout",
+		cfg.POET.RequestTimeout, "default timeout for poet requests")
 
 	/**======================== bootstrap data updater Flags ========================== **/