+++ config/presets/standalone.go
@@ -80,10 +80,10 @@ func standalone() config.Config {
 	conf.POET.GracePeriod = 12 * time.Second
 	conf.POET.CycleGap = 30 * time.Second
 	conf.POET.PhaseShift = 30 * time.Second
-	conf.POET.GetProofTimeout = 12 * time.Second // equal to DefaultRequestTimeout
+	conf.POET.FetchProofTimeout = 12 * time.Second // equal to RequestTimeout
 	conf.POET.SubmitChallengeTimeout = 12 * time.Second
-	// DefaultRequestTimeout = RequestRetryDelay * 2 * MaxRequestRetries*(MaxRequestRetries+1)/2
-	conf.POET.DefaultRequestTimeout = 12 * time.Second
+	// RequestTimeout = RequestRetryDelay * 2 * MaxRequestRetries*(MaxRequestRetries+1)/2
+	conf.POET.RequestTimeout = 12 * time.Second
 	conf.POET.RequestRetryDelay = 1 * time.Second
 	conf.POET.MaxRequestRetries = 3