+++ systest/tests/distributed_post_verification_test.go
@@ -64,8 +64,8 @@ func TestPostMalfeasanceProof(t *testing.T) {
 	cfg.P2P.DataDir = filepath.Join(testDir, "p2p-dir")
 	require.NoError(t, os.Mkdir(cfg.P2P.DataDir, os.ModePerm))
 
-	cfg.POET.DefaultRequestTimeout = time.Minute
-	cfg.POET.GetProofTimeout = time.Minute
+	cfg.POET.RequestTimeout = time.Minute
+	cfg.POET.FetchProofTimeout = time.Minute
 	cfg.POET.SubmitChallengeTimeout = time.Minute
 	cfg.POET.MaxRequestRetries = 10